# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AffordabilityOut": {
        "properties": {
          "allowance": {
            "title": "Allowance",
            "type": "number"
          },
          "depositEur": {
            "title": "Depositeur",
            "type": "number"
          },
          "incomeEur": {
            "title": "Incomeeur",
            "type": "number"
          },
          "limitedBy": {
            "enum": [
              "income",
              "deposit"
            ],
            "title": "Limitedby",
            "type": "string"
          },
          "loanEur": {
            "title": "Loaneur",
            "type": "number"
          },
          "loanToIncome": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loantoincome"
          },
          "loanToValue": {
            "title": "Loantovalue",
            "type": "number"
          },
          "maxLoanByIncomeEur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxloanbyincomeeur"
          },
          "maxPriceByDepositEur": {
            "title": "Maxpricebydepositeur",
            "type": "number"
          },
          "maxPriceEur": {
            "title": "Maxpriceeur",
            "type": "number"
          },
          "measuresNote": {
            "title": "Measuresnote",
            "type": "string"
          },
          "monthlyRepaymentEur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthlyrepaymenteur"
          },
          "note": {
            "default": "Information only, not financial or tax advice. Lenders decide individually, and your own circumstances or reliefs may change the figures.",
            "title": "Note",
            "type": "string"
          },
          "rulesVersion": {
            "title": "Rulesversion",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "stampDutyEur": {
            "title": "Stampdutyeur",
            "type": "number"
          }
        },
        "required": [
          "rulesVersion",
          "sources",
          "incomeEur",
          "depositEur",
          "loanToIncome",
          "loanToValue",
          "maxLoanByIncomeEur",
          "maxPriceByDepositEur",
          "maxPriceEur",
          "loanEur",
          "limitedBy",
          "monthlyRepaymentEur",
          "stampDutyEur",
          "allowance",
          "measuresNote"
        ],
        "title": "AffordabilityOut",
        "type": "object"
      },
...
      "DutyBandOut": {
        "properties": {
          "dutyEur": {
            "title": "Dutyeur",
            "type": "number"
          },
          "fromEur": {
            "title": "Fromeur",
            "type": "number"
          },
          "rate": {
            "title": "Rate",
            "type": "number"
          },
          "toEur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toeur"
          }
        },
        "required": [
          "fromEur",
          "toEur",
          "rate",
          "dutyEur"
        ],
        "title": "DutyBandOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Source": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "verifiedOn": {
            "format": "date",
            "title": "Verifiedon",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "verifiedOn"
        ],
        "title": "Source",
        "type": "object"
      },
...
      "StampDutyOut": {
        "properties": {
          "bands": {
            "items": {
              "$ref": "#/components/schemas/DutyBandOut"
            },
            "title": "Bands",
            "type": "array"
          },
          "considerationEur": {
            "title": "Considerationeur",
            "type": "number"
          },
          "dutyEur": {
            "title": "Dutyeur",
            "type": "number"
          },
          "effectiveRate": {
            "title": "Effectiverate",
            "type": "number"
          },
          "notCovered": {
            "title": "Notcovered",
            "type": "string"
          },
          "note": {
            "default": "Information only, not financial or tax advice. Lenders decide individually, and your own circumstances or reliefs may change the figures.",
            "title": "Note",
            "type": "string"
          },
          "priceEur": {
            "title": "Priceeur",
            "type": "number"
          },
          "rulesVersion": {
            "title": "Rulesversion",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "vatEur": {
            "title": "Vateur",
            "type": "number"
          },
          "vatRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vatrate"
          }
        },
        "required": [
          "rulesVersion",
          "sources",
          "priceEur",
          "considerationEur",
          "vatEur",
          "vatRate",
          "dutyEur",
          "effectiveRate",
          "bands",
          "notCovered"
        ],
        "title": "StampDutyOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Ireland's Property Price Register, made usable. Contains Residential Property Price Register data \u00a9 Property Services Regulatory Authority; it may contain errors.",
    "title": "PPR Map API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tools/affordability": {
      "get": {
        "description": "The most the Central Bank measures allow you to borrow and spend: the lower of the\nincome limit and the deposit limit. Lenders may lend less.",
        "operationId": "affordability_api_v1_tools_affordability_get",
        "parameters": [
          {
            "in": "query",
            "name": "grossIncome",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100000000.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ],
              "title": "Grossincome"
            }
          },
          {
            "in": "query",
            "name": "deposit",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100000000.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ],
              "title": "Deposit"
            }
          },
          {
            "in": "query",
            "name": "secondIncome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100000000.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ],
              "default": "0",
              "title": "Secondincome"
            }
          },
          {
            "in": "query",
            "name": "buyer",
            "required": false,
            "schema": {
              "default": "first_time_buyer",
              "enum": [
                "first_time_buyer",
                "second_and_subsequent",
                "buy_to_let"
              ],
              "title": "Buyer",
              "type": "string"
            }
          },
          {
            "description": "Shorthand for buyer",
            "in": "query",
            "name": "firstTimeBuyer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Shorthand for buyer",
              "title": "Firsttimebuyer"
            }
          },
          {
            "in": "query",
            "name": "termYears",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 35,
              "minimum": 5,
              "title": "Termyears",
              "type": "integer"
            }
          },
          {
            "description": "Your quoted rate",
            "in": "query",
            "name": "ratePct",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 20.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Your quoted rate",
              "title": "Ratepct"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffordabilityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Affordability",
        "tags": [
          "tools"
        ]
      }
    },
...
    "/api/v1/tools/stamp-duty": {
      "get": {
        "description": "Stamp duty on one home at today's rates. On a new home it is charged on the price\nwithout VAT.",
        "operationId": "stamp_duty_api_v1_tools_stamp_duty_get",
        "parameters": [
          {
            "description": "Price in euro",
            "in": "query",
            "name": "price",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0.0,
                  "maximum": 100000000.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ],
              "description": "Price in euro",
              "title": "Price"
            }
          },
          {
            "in": "query",
            "name": "isNew",
            "required": false,
            "schema": {
              "default": false,
              "title": "Isnew",
              "type": "boolean"
            }
          },
          {
            "description": "For a new home: the price includes VAT",
            "in": "query",
            "name": "vatInclusive",
            "required": false,
            "schema": {
              "default": true,
              "description": "For a new home: the price includes VAT",
              "title": "Vatinclusive",
              "type": "boolean"
            }
          },
          {
            "description": "A new apartment in a block of 3+ with shared access (9% VAT since 8 Oct 2025)",
            "in": "query",
            "name": "qualifyingApartment",
            "required": false,
            "schema": {
              "default": false,
              "description": "A new apartment in a block of 3+ with shared access (9% VAT since 8 Oct 2025)",
              "title": "Qualifyingapartment",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampDutyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stamp Duty",
        "tags": [
          "tools"
        ]
      }
    }
  }
}